# Conflicts:
#	scripts/sample_app5/config/applications.xml
#	scripts/sample_app5/config/edge_devices.xml
#	src/edu/boun/edgecloudsim/core/SimSettings.java
#	src/edu/boun/edgecloudsim/mobility/StaticRangeMobility.java
#	src/edu/boun/edgecloudsim/storage/ObjectGenerator.java
#	src/edu/boun/edgecloudsim/task_generator/IdleActiveStorageLoadGenerator.java

+++ src/edu/boun/edgecloudsim/storage/RedisListHandler.java
@@ -1,5 +1,3 @@
-//TODO: Create redis connection exceptions
-//TODO: Add constants (PARITY,DATA,MD) to parity (object:md*)
 package edu.boun.edgecloudsim.storage;
 
 import edu.boun.edgecloudsim.core.SimSettings;
@@ -17,6 +15,7 @@ import redis.clients.jedis.util.Slowlog;
 import javax.swing.*;
 import java.io.BufferedWriter;
 import java.io.File;
+import java.nio.file.*;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.util.*;
@@ -62,6 +61,12 @@ public class RedisListHandler {
 
     //Generate list of all object locations in the system for orchestration
     private static void listObjectInSystem(ObjectGenerator OG) throws IOException {
+        Path pLoc = Paths.get("/tmp/Object_Locations.txt");
+        Path pDist = Paths.get("/tmp/OBJECT_DISTRIBUTION.txt");
+        if(Files.exists(pLoc) && Files.exists(pDist)){
+            SimLogger.print("Object locations and distribution files exist"+"\n");
+            return;
+        }
         File objectFile = new File("/tmp/Object_Locations.txt");
         File objectDistFile = new File("/tmp/OBJECT_DISTRIBUTION.txt");
         FileWriter objectFW = new FileWriter(objectFile, false),
@@ -71,9 +76,11 @@ public class RedisListHandler {
 
 
         objectBW.write("object,locations\n");
-        objectDistBW.write("Object Name;Object Type;Occurrences");
+        objectDistBW.write("Object Name,Object Type,Occurrences");
         objectDistBW.newLine();
         for (Map<String,String> KV : OG.getListOfObjects()) {
+            if(KV.get("type").equals("metadata"))
+                continue;
             objectBW.write("object:" + KV.get("id")+","+KV.get("locations"));
             objectBW.newLine();
 
@@ -98,6 +105,8 @@ public class RedisListHandler {
         OG = new ObjectGenerator(objectPlacementPolicy);
         //Generate Redis objects for this host
         for (Map<String,String> KV : OG.getListOfObjects()) {
+            if(KV.get("type").equals("metadata"))
+                continue;
             String locations = KV.get("locations");
             StringTokenizer st = new StringTokenizer(locations, " "); // Space as delimiter
             Set<String> locationsSet = new HashSet<String>();
@@ -192,8 +201,10 @@ public class RedisListHandler {
     //Returns data object IDs in index 0 and parity in index 1
     public static String[] getStripeObjects(String metadataID){
         Jedis jedis = new Jedis(localhost, 6379);
-        String dataObjects = jedis.hget("object:"+metadataID,"data");
-        String parityObjects = jedis.hget("object:"+metadataID,"parity");
+        if (!metadataID.contains("object:"))
+            metadataID = "object:"+metadataID;
+        String dataObjects = jedis.hget(metadataID,"data");
+        String parityObjects = jedis.hget(metadataID,"parity");
         jedis.close();
         return new String[] {dataObjects,parityObjects};
